fix: remove ValidatedString and TestrunOrSkipped
I previously added these because we were having issues storing very long test
names and I wanted to "validate" the data that came in to make sure that we're
not including random strings for test names that might include source code. The
main problem was that we weren't able to index that data in timescale and that
was causing issues. I've since removed those indexes and if we have to index
on names we'll have to use full text indexes next time, but for now we don't
need this anymore. In the future, we'll likely have to add some mechanism for
users to "ignore" irrelevant tests in the UI, or filter the names of tests. In
any case, the scope of this library is to simply parse a JUnit XML file, so we
should stick to that here, and avoid trying to do more.